Module rsp2_minimize::cg

Re-exports

pub use self::stop_condition::StopCondition;

Modules

hager_beta

settings
Please do not manually construct anything in here. None of this is future-safe.

stop_condition

Structs

AlgorithmState

Builder
Provides a significantly greater level of customization for performing conjugate gradient.

Failure
An error type extended with some additional data.

Output

Enums

AcgsdError

Traits

BuildAlgorithmStateFn
Trait for producing fresh instances of an AlgorithmStateFn.

DiffFn
Represents an objective function with an analytically known gradient, to be minimized by conjugate gradient.

Functions

acgsd
Perform conjugate gradient using the default configuration for ACGSD, and with a stop condition that can be deserialized from JSON.

cg_descent
Perform conjugate gradient using the default configuration for CG-DESCENT, and with a stop condition that can be deserialized from JSON.

get_basic_output_fn

output_fn_cosine_formatter
Helper for producing a column with the cosines between succesive directions, just like the one produced by basic_output_fn.

output_fn_dv_formatter
Helper for producing a column with the change in value between succesive iterations, just like the one produced by basic_output_fn.