Type Definition rsp2_minimize::cg::stop_condition::StopCondition

type StopCondition = Cereal<Simple>;

Configuration for the built-in implementation of stop conditions.

The recommended method for constructing one of these is to deserialize it from JSON. The set of basic conditions is documented in #[serde] annotations on Simple, and there are additionally "all" and "any" meta-conditions that are provided by crate::stop_condition::Cereal.

If this doesn’t fit your needs, you can use CG’s Builder API to define your own arbitrary stop conditions as functions of AlgorithmState.

Implementations

impl StopCondition

pub fn to_function(&self) -> impl Clone + FnMut(AlgorithmState<'_>) -> bool

Convert to the more general form accepted by the Builder API.