Trait rsp2_minimize::cg::DiffFn

pub trait DiffFn {
    type Error;
    fn compute(&mut self, pos: &[f64]) -> Result<(f64, Vec<f64>), Self::Error>;

    fn check(&mut self, _pos: &[f64]) -> Result<(), Self::Error> { ... }
}

Represents an objective function with an analytically known gradient, to be minimized by conjugate gradient.

You can use a fn(&[f64]) -> Result<(f64, Vec<f64>), E> closure, but note that you must annotate the closure argument enough for Rust to see the higher-ranked lifetime. (i.e. |pos: &_| ...).

Associated Types

type Error

Loading content...
Required methods

fn compute(&mut self, pos: &[f64]) -> Result<(f64, Vec<f64>), Self::Error>

Compute the value and gradient at a position.

Loading content...
Provided methods

fn check(&mut self, _pos: &[f64]) -> Result<(), Self::Error>

This is called at the end of each line search.

This gives the DiffFn an opportunity to check for properties that are required to hold for any “realistic” position. (whereas compute may frequently be called on absurd positions that are speculatively produced by linesearch)

Loading content...