Module rsp2_minimize::stop_condition

Exposes a serializable expression language for algorithmic stop conditions.

Modules

prelude

Structs

Rpn
A composite stop condition expressed in reverse-Polish notation.

Enums

Cereal
Type that stop condition config can deserialize directly into.

LogicalExpression
Represents logical expressions of T in JSON.

Traits

ShouldStop
Generic trait for stop conditions.